test: split merged api test modules and remove F811 ignore
